+++ Src/Results/IResult{T}.cs
@@ -1,4 +1,8 @@
-﻿using System.Runtime.Serialization;
+﻿// <copyright file="IResult{T}.cs" company="Zentient Framework Team">
+// Copyright © 2025 Zentient Framework Team. All rights reserved.
+// </copyright>
+
+using System.Runtime.Serialization;
 using System.Text.Json.Serialization;
 
 #pragma warning disable CA1715 // Type names should not have the same name as a namespace
@@ -96,4 +100,3 @@ namespace Zentient.Results
         U Match<U>(Func<T, U> onSuccess, Func<IReadOnlyList<ErrorInfo>, U> onFailure);
     }
 }
-