+++ Src/Utilities/Guard.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿// <copyright file="Guard.cs" company="Zentient Framework Team">
+// Copyright © 2025 Zentient Framework Team. All rights reserved.
+// </copyright>
+
+using System;
 using System.Collections.Generic;
 
 using Zentient.Utilities;